% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{blob}
\alias{blob}
\title{BLOB}
\usage{
blob(dir, ext = "[.]pdf$", n = 10000L)
}
\arguments{
\item{dir}{A string of the directory}

\item{ext}{A string of the file extension(s)}

\item{n}{An integer of the (maximal) number of records to be read.}
}
\value{
A tibble with the columns FileName and BLOB
}
\description{
BLOB
}
\seealso{
readBin
}
