% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epsg}
\alias{epsg}
\title{epsg}
\format{
A data frame containing the European Petroleum Survey Group
(EPSG) geodetic parameter dataset as distributed with PROJ.4 software
(previously supplied by rgdal). Equivalent Proj4 strings as well as other
information on coordinate systems can be found at https://spatialreference.org.
}
\usage{
epsg
}
\description{
epsg
}
\examples{
epsg
}
\keyword{datasets}
