% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{ps_activate_sfc}
\alias{ps_activate_sfc}
\title{Activate sfc column}
\usage{
ps_activate_sfc(x, sfc_name = "geometry")
}
\arguments{
\item{x}{The object object.}

\item{sfc_name}{A string indicating the name of the sfc column to activate.}
}
\description{
Activates sfc column.
}
