% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{ps_active_sfc_name}
\alias{ps_active_sfc_name}
\title{Name of active sfc column}
\usage{
ps_active_sfc_name(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A character vector of the name of the active sfc column.
}
\description{
Name of active sfc column
}
