% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ps_batch_transform}
\alias{ps_batch_transform}
\title{Batch transform}
\usage{
ps_batch_transform(in_dir, out_dir, recursive = T, in_pattern = "[.]shp$",
  out_extension = ".sqlite", in_crs, out_crs)
}
\arguments{
\item{in_dir}{A string of the path to the directory containing spatial files.}

\item{out_dir}{A string of the path to the directory that output files will be saved to.}

\item{recursive}{A logical specifying whether to search directory recursively.}

\item{in_pattern}{A string of the pattern to use when searching for input files.}

\item{out_extension}{A string of the extension specifying which file format output files will be saved to.}

\item{in_crs}{An integer of the EPSG that files are currently in (only set if crs not currently already present).}

\item{out_crs}{An integer of the EPSG that files will be transformed to.}
}
\value{
Files with specified crs and extension written to output directory.
}
\description{
Batch transform spatial files. Function finds all files with specified extension pattern within input directory
and sets crs if does not already exist. Files are projected to new crs and written
to output directory with specified extension and file format (allows any extension recognised by sf::st_write function).
}
