% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_bbox_ggmap}
\alias{ps_bbox_ggmap}
\title{Get clipped basemap}
\usage{
ps_bbox_ggmap(x, source, maptype)
}
\arguments{
\item{x}{A bbox object (or numeric vector of length 4 indicating xmin, ymin, xmax, ymax). Must be long/lat.}

\item{source}{Google Maps ("google"), OpenStreetMap ("osm"), Stamen Maps ("stamen"), or CloudMade maps ("cloudmade")}

\item{maptype}{Character string providing map theme. Options available are "terrain", "terrain-background",
"satellite", "roadmap", and "hybrid" (google maps), "terrain", "watercolor", and "toner" (stamen maps), or a positive integer for cloudmade maps (see ?get_cloudmademap)}
}
\value{
A ggmap object
}
\description{
Get ggmap basemap from bbox
}
