% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_blob}
\alias{ps_blob}
\title{BLOB}
\usage{
ps_blob(dir, ext = "[.]pdf$", n = 10000L, recursive = TRUE)
}
\arguments{
\item{dir}{A string of the directory}

\item{ext}{A string of the file extension(s)}

\item{n}{An integer of the (maximal) number of records to be read.}

\item{recursive}{A flag indicating whether no recurse into subdirectories.}
}
\value{
A tibble with the columns FileName and BLOB
}
\description{
BLOB
}
\seealso{
readBin
}
