% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_coords_to_sf}
\alias{ps_coords_to_sf}
\title{Convert point coordinates to sf (active geometry) column.}
\usage{
ps_coords_to_sf(x, coords = c("X", "Y"), crs = 4326,
  new_name = "geometry")
}
\arguments{
\item{x}{The object with columns}

\item{coords}{A character vector of specifying the two columns with the point information.}

\item{crs}{An integer with the EPSG code, or character with proj4string.}

\item{new_name}{A string of the name of the sf column.}
}
\value{
The modified object with the coordinates removed
}
\description{
Convert point coordinates to sf (active geometry) column.
}
