% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_coords_to_sfc}
\alias{ps_coords_to_sfc}
\title{Convert point coordinates to sfc (inactive geometry) column.}
\usage{
ps_coords_to_sfc(x, coords = c("X", "Y"), crs = 4326,
  sfc_name = "geometry")
}
\arguments{
\item{x}{The object with columns}

\item{coords}{A character vector of specifying the two columns with the point information.}

\item{crs}{An integer with the EPSG code, or character with proj4string.}

\item{sfc_name}{A string of the name of the sfc column.}
}
\value{
The modified object with the coordinates removed
}
\description{
Convert point coordinates to sfc (inactive geometry) column.
}
