% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_create_bounds}
\alias{ps_create_bounds}
\title{Create bounds}
\usage{
ps_create_bounds(x, pad)
}
\arguments{
\item{x}{A sf object. If long/lat, object is temporarily transformed to UTM to add pad distance.}

\item{pad}{A numeric vector indicating amount in metres to pad bounds.
If vector of length 1, amount is added to all sides;
if vector of length 4 amount is added to left, top, right and bottom, respectively.}
}
\value{
sfc polygon
}
\description{
Create a sfc polygon rectangle as padded bbox of sf object.
}
