% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{ps_dd2ddm}
\alias{ps_dd2ddm}
\title{Decimal Degrees to Degrees and Decimal Minutes.}
\usage{
ps_dd2ddm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.
}
\description{
Decimal Degrees to Degrees and Decimal Minutes.
}
\examples{
ps_dd2ddm(c(70.654, NA, -56.654))
}
