% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{ps_dd2dmds}
\alias{ps_dd2dmds}
\title{Decimal Degrees to Degrees and Decimal minutes.}
\usage{
ps_dd2dmds(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.
}
\description{
Decimal Degrees to Degrees and Decimal minutes.
}
\examples{
ps_dd2dmds(c(70.654, NA, -56.654))
}
