% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{ps_ddm2dd}
\alias{ps_ddm2dd}
\title{Degrees and Decimal Minutes to Decimal Degrees}
\usage{
ps_ddm2dd(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A numeric vector.
}
\description{
Degrees and Decimal Minutes to Decimal Degrees
}
\examples{
ps_ddm2dd(c("70 39.24", NA, "-56 39.24"))
}
