% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_deactivate_sf}
\alias{ps_deactivate_sf}
\title{Deactivate sf column}
\usage{
ps_deactivate_sf(x)
}
\arguments{
\item{x}{A sf object.}
}
\description{
Deactives geometry column ('sf column').
}
