% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{ps_deactivate_sfc}
\alias{ps_deactivate_sfc}
\title{Deactivate active sf column}
\usage{
ps_deactivate_sfc(x)
}
\arguments{
\item{x}{A sf object.}
}
\description{
Deactives active geometry column ('sf column').
}
