% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{ps_dm2mds}
\alias{ps_dm2mds}
\title{Decimal Minutes to Minutes and Decimal Seconds.}
\usage{
ps_dm2mds(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector.
}
\description{
Decimal Minutes to Minutes and Decimal Seconds.
}
\examples{
ps_dm2mds(c(10.654, NA))
}
