% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{ps_elevation_google}
\alias{ps_elevation_google}
\title{Add Z column to sfc based on google maps}
\usage{
ps_elevation_google(
  x,
  sfc_name = ps_active_sfc_name(x),
  key = Sys.getenv("GOOGLE_MAPS_ELEVATION_API_KEY")
)
}
\arguments{
\item{x}{The sf object.}

\item{sfc_name}{A string of the sfc column name.}

\item{key}{The string of the google maps elevation api key.}
}
\value{
The sf object with a Z column of the elevation in metres.
}
\description{
Add Z column to sfc based on google maps
}
