% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{ps_equal_crs}
\alias{ps_equal_crs}
\title{Check that sfcs have same crs}
\usage{
ps_equal_crs(x, sfc_names = ps_sfc_names(x))
}
\arguments{
\item{x}{The object}

\item{sfc_names}{A character vector of the sfc column names}
}
\value{
Logical
}
\description{
Check that sfcs have same crs
}
