% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ps_fwa_gdbs}
\alias{ps_fwa_gdbs}
\title{FWA gdbs}
\usage{
ps_fwa_gdbs(dir = "~/Poisson/Data/spatial/fwa/gdb")
}
\arguments{
\item{dir}{A character string indicating path to directory holding fwa geodatabases.}
}
\value{
A factor of the geodatabase names.
}
\description{
FWA gdbs
}
