% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ps_fwa_shortcuts}
\alias{ps_fwa_shortcuts}
\title{FWA shortcuts}
\usage{
ps_fwa_shortcuts(gdb = "FWA_BC.gdb",
  dir = "~/Poisson/Data/spatial/fwa/gdb/")
}
\arguments{
\item{gdb}{A character string vector indicating FWA geodatabases to extract layer shortcuts from.}

\item{dir}{A character string indicating path to directory holding fwa geodatabases.}
}
\value{
A factor of the layer names within specified geodatabase.
}
\description{
FWA shortcuts
}
