% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_ggmap_to_raster}
\alias{ps_ggmap_to_raster}
\title{Convert ggmap to raster}
\usage{
ps_ggmap_to_raster(x)
}
\arguments{
\item{x}{A ggmap object.}
}
\value{
raster object.
}
\description{
Convert ggmap to raster
}
