% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{ps_inactive_sfc_names}
\alias{ps_inactive_sfc_names}
\title{Name of inactive sfc column(s)}
\usage{
ps_inactive_sfc_names(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A character vector of the name of the inactive sfc column(s).
}
\description{
Name of inactive sfc column(s)
}
