% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ps_load_gdb}
\alias{ps_load_gdb}
\title{Load spatial files}
\usage{
ps_load_gdb(dir = "~/Poisson/Data/spatial/fwa/gdb/FWA_BC.gdb",
  layers = NULL, crs = NULL, rename = identity, envir = parent.frame(),
  fun = identity, ...)
}
\arguments{
\item{dir}{A string of the geodatabase directory.}

\item{layers}{A character string vector indicating which layers to load.}

\item{crs}{Default crs of layer if missing.}

\item{rename}{A function that is used to rename files (after removing extension .csv) before they are passed to make.names.}

\item{envir}{The environment to assign the data frames.}

\item{fun}{A function that is applied to all sf objects before they are assigned to envir.}

\item{...}{Additional arguments passed to \code{st_read}.}
}
\value{
An invisible character vector of the layer names.
}
\description{
Load spatial files
}
