% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ps_load_spatial}
\alias{ps_load_spatial}
\title{Load spatial files}
\usage{
ps_load_spatial(dir = ".", pattern = NULL, recursive = FALSE,
  crs = NULL, rename = identity, envir = parent.frame(),
  fun = identity, ...)
}
\arguments{
\item{dir}{A string of the directory.}

\item{pattern}{An optional regular expression. Only file names which match the regular expression will be returned.
If NULL, all files with extensions readable by st_read wlll be returned.}

\item{recursive}{A flag indicating whether to include files in subdirectories.}

\item{crs}{Default crs of object if missing.}

\item{rename}{A function that is used to rename files (after removing extension .csv) before they are passed to make.names.}

\item{envir}{The environment to assign the data frames.}

\item{fun}{A function that is applied to all sf objects before they are assigned to envir.}

\item{...}{Additional arguments passed to \code{st_read}.}
}
\value{
An invisible character vector of the file names.
}
\description{
Load spatial files
}
