% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlat.R
\name{ps_longlat_to_sf}
\alias{ps_longlat_to_sf}
\title{Convert Longitude and Latitude coordinates to sf (active geometry) column.}
\usage{
ps_longlat_to_sf(x, sf_name = "geometry")
}
\arguments{
\item{x}{The object with columns Latitude and Longitude.}

\item{sf_name}{A string of the name of the sf column.}
}
\value{
The modified object with Longitude and Latitude removed
}
\description{
Assumes Longitude and Latitude are in WGS84.
}
