% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlat.R
\name{ps_longlat_to_sfc}
\alias{ps_longlat_to_sfc}
\title{Convert Longitude and Latitude coordinates to sfc column.}
\usage{
ps_longlat_to_sfc(x, sfc_name = "geometry", activate = TRUE)
}
\arguments{
\item{x}{The object with columns Latitude and Longitude.}

\item{sfc_name}{A string of the name of the sfc column to create.}

\item{activate}{A flag indicating whether to activate the sfc.
#' @param retain_orig A flag indicating if the input coordinates should be retained as columns in the dataframe.}
}
\value{
The modified object with Longitude and Latitude removed
}
\description{
Assumes Longitude and Latitude are in WGS84.
}
