% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat.R
\name{ps_lonlat_to_sfc}
\alias{ps_lonlat_to_sfc}
\title{Convert Longitude and Latitude coordinates to sfc (inactive geometry) column.}
\usage{
ps_lonlat_to_sfc(x, sfc_name = "geometry")
}
\arguments{
\item{x}{The object with columns Latitude and Longitude.}

\item{sfc_name}{A string of the name of the sf column.}
}
\value{
The modified object with Longitude and Latitude removed
}
\description{
Assumes Longitude and Latitude are in WGS84.
}
