% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{ps_nearest}
\alias{ps_nearest}
\title{Nearest Neighbour}
\usage{
ps_nearest(x, y, by = c("X", "Y"), dist_col = NULL, keep_names = FALSE, ...)
}
\arguments{
\item{x}{A data.frame, tibble or sf object.}

\item{y}{A data.frame, tibble or sf object.}

\item{by}{A possibly named character vector specifying the column(s) to calculate the distance over.}

\item{dist_col}{A string indicating the name of the column to save the distance in.}

\item{keep_names}{A flag specifying whether to keep the original sfc names instead of renaming them to geometry.}

\item{...}{Not used}
}
\description{
Each row in object x is bound with its closest neighbour in object y.
Uses the nabor package.
}
\details{
The column(s) to use when calculating the distances are converted to numeric values.
Missing values are currently not permitted.

By default sf objects have their sf (active geometry) column renamed to geometry. The nearest calculation for non-point sf objects is based on feature vertices.
}
