% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{ps_nearest1}
\alias{ps_nearest1}
\title{Nearest Row}
\usage{
ps_nearest1(x, y, by)
}
\arguments{
\item{x}{A data.frame with the column by}

\item{y}{A data.frame with the column by.}

\item{by}{The unquoted name of the column to calculate the inner join by.}
}
\value{
A tibble of x with the columns from y and a column Distance indicating the distance between the by values in x and y.
}
\description{
Joins two data.frames based on the nearest value in by.
}
