% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tracks-gpx.R
\name{ps_read_tracks_gpxs}
\alias{ps_read_tracks_gpxs}
\title{Read GPX}
\usage{
ps_read_tracks_gpxs(
  dir,
  pattern = "[.]gpx$",
  recursive = FALSE,
  tz = getOption("ps.tz", "UTC"),
  crs = getOption("ps.crs", 4326)
)
}
\arguments{
\item{dir}{A string of the directory name.}

\item{pattern}{A string of the pattern to use when searching for files.}

\item{recursive}{A flag indicating whether to recurse into subdirectories.}

\item{tz}{A string of the tz for the date times.}

\item{crs}{An integer of the EPSG for the projection of the points.}
}
\value{
An sf object with a tibble of the datetime (POSIXct) and the file path (character) and a sfc_POINT geometry of three dimensional points where the third dimension is the elevation in m.
}
\description{
Reads tracks from a gpx file.
}
\seealso{
ps_read_tracks_gpx
}
