% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-waypoints-gpx.R
\name{ps_read_waypoints_gpx}
\alias{ps_read_waypoints_gpx}
\title{Read GPX}
\usage{
ps_read_waypoints_gpx(file, tz = getOption("ps.tz", "UTC"),
  crs = getOption("ps.crs", 4326))
}
\arguments{
\item{file}{A string of the path to the file.}

\item{tz}{A string of the tz for the date times.}

\item{crs}{An integer of the EPSG for the projection of the points.}
}
\value{
An sf object with a tibble of the datetime and a sfc_POINT geometry of three dimensional points where the third dimension is the elevation in m.
}
\description{
Reads waypoints from a gpx file.
}
\seealso{
ps_read_waypoints_gpxs
}
