% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_remove_any_sfc}
\alias{ps_remove_any_sfc}
\title{Remove any sfc column}
\usage{
ps_remove_any_sfc(x)
}
\arguments{
\item{x}{A sf object.}
}
\description{
All sfc columns are removed. Active sfc column ('sf column') is not removed.
}
