% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_remove_sfc}
\alias{ps_remove_sfc}
\title{Remove sfc column}
\usage{
ps_remove_sfc(x, sfc_column)
}
\arguments{
\item{x}{A sf object.}

\item{sfc_column}{A string indicating the name of the sfc column(s) to remove.}
}
\description{
Remove sfc column(s). Active sfc column ('sf column') will not be removed even if specified.
}
