% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{ps_remove_sfcs}
\alias{ps_remove_sfcs}
\title{Remove sfc columns}
\usage{
ps_remove_sfcs(x, sfc_names = ps_sfc_names(x))
}
\arguments{
\item{x}{An sf object.}

\item{sfc_names}{A character vector indicating the name of the sfc column(s) to remove.}
}
\description{
Remove specified sfc (geometry) column(s).
}
