% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{ps_rename_active_sfc}
\alias{ps_rename_active_sfc}
\title{Rename active sfc column}
\usage{
ps_rename_active_sfc(x, new_name = "geometry")
}
\arguments{
\item{x}{The object.}

\item{new_name}{A string of the new column name.}
}
\value{
The modified object
}
\description{
Renames the active sfc column.
}
