% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_rename_sf}
\alias{ps_rename_sf}
\title{Rename sf column}
\usage{
ps_rename_sf(x, sf_name = "geometry")
}
\arguments{
\item{x}{The object.}

\item{sf_name}{A string of the new column name.}
}
\value{
The modified object
}
\description{
Renames the active geometry column.
}
