% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ps_rename_sf_column_name}
\alias{ps_rename_sf_column_name}
\title{Rename geometry column}
\usage{
ps_rename_sf_column_name(x, new_name = "geometry")
}
\arguments{
\item{x}{The object.}

\item{new_name}{A string of the new column name.}
}
\value{
The modified object
}
\description{
Rename geometry column
}
