% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_set_sf}
\alias{ps_set_sf}
\title{Set sf column}
\usage{
ps_set_sf(x, sfc_name = "geometry")
}
\arguments{
\item{x}{A sf object.}

\item{sfc_name}{A string indicating the name of the sf column to set.}
}
\description{
Set active geometry column ('sf column').
}
