% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ps_sf_column_name}
\alias{ps_sf_column_name}
\title{Name of geometry column(s)}
\usage{
ps_sf_column_name(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A character vector of the name of the geometry column(s).
}
\description{
Name of geometry column(s)
}
