% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_sf_ggmap}
\alias{ps_sf_ggmap}
\title{Ggmap from sf}
\usage{
ps_sf_ggmap(x, pad, source, maptype)
}
\arguments{
\item{x}{A sf object.}

\item{pad}{A numeric vector indicating amount in metres to pad bbox. If vector of length 1, that amount is added to all sides; if vector of length 4 amount is added to left, top, right and bottom, respectively.}

\item{source}{Google Maps ("google"), OpenStreetMap ("osm"), Stamen Maps ("stamen"), or CloudMade maps ("cloudmade").}

\item{maptype}{Character string providing map theme. Options available are "terrain", "terrain-background", "satellite", "roadmap", and "hybrid" (google maps),
"terrain", "watercolor", and "toner" (stamen maps),
or a positive integer for cloudmade maps (see ?get_cloudmademap).}
}
\value{
ggmap object.
}
\description{
Get ggmap basemap from (padded) sf object.
}
