% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_sf_ggmap_df}
\alias{ps_sf_ggmap_df}
\title{Ggmap from sf to data.frame}
\usage{
ps_sf_ggmap_df(x, pad, source, maptype)
}
\arguments{
\item{x}{A sf object}

\item{pad}{A numeric vector indicating amount in metres to pad bbox. If vector of length 1, that amount is added to all sides; if vector of length 4 amount is added to left, top, right and bottom, respectively.}

\item{source}{Google Maps ("google"), OpenStreetMap ("osm"), Stamen Maps ("stamen"), or CloudMade maps ("cloudmade")}

\item{maptype}{Character string providing map theme. Options available are "terrain", "terrain-background", "satellite", "roadmap", and "hybrid" (google maps), "terrain", "watercolor", and "toner" (stamen maps), or a positive integer for cloudmade maps (see ?get_cloudmademap)}
}
\value{
A data.frame which may be used to plot basemap with ggplot2 using: geom_point(data = x, aes(x = x, y = y, col = rgb(layer.1/255, layer.2/255, layer.3/255)))
}
\description{
A wrapper function to quickly get plottable ggmap basemap clipped to padded bbox of sf object
}
