% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_sf_name}
\alias{ps_sf_name}
\title{Name of sf column}
\usage{
ps_sf_name(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A character vector of the name of the active geometry column ('sf column').
}
\description{
Name of sf column
}
