% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_sf_rectangle}
\alias{ps_sf_rectangle}
\title{Create sf rectangle}
\usage{
ps_sf_rectangle(x, crs)
}
\arguments{
\item{x}{A bbox object (or numeric vector of length 4 indicating xmin, ymin, xmax, ymax).}

\item{crs}{coordinate reference system: integer with the EPSG code, or character with proj4string}
}
\value{
sfc polygon.
}
\description{
Create sf rectangle
}
