% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_sf_to_coords}
\alias{ps_sf_to_coords}
\title{Convert sf (active geometry) to pair of coordinates column.}
\usage{
ps_sf_to_coords(x, X = "X", Y = "Y")
}
\arguments{
\item{x}{The object with columns}

\item{X}{A string of the name of the X coordinate.}

\item{Y}{A string of the name of the Y coordinate.}
}
\value{
A tibble with the sf column removed
}
\description{
Convert sf (active geometry) to pair of coordinates column.
}
