% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat.R
\name{ps_sf_to_lonlat}
\alias{ps_sf_to_lonlat}
\title{Convert sf (active geometry) column to Longitude and Latitude in WGS84.}
\usage{
ps_sf_to_lonlat(x)
}
\arguments{
\item{x}{The sf object.}
}
\value{
A tibble with Longitude and Latitude.
}
\description{
Convert sf (active geometry) column to Longitude and Latitude in WGS84.
}
