% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-z.R
\name{ps_sfc_add_z}
\alias{ps_sfc_add_z}
\title{Add elevation (Z) to sf points.}
\usage{
ps_sfc_add_z(x, sfc_column = ps_active_sfc_name(x), z_column = "Elevation",
  new_column = NULL, remove_z = TRUE)
}
\arguments{
\item{x}{A sf object.}

\item{sfc_column}{A string of the name of the sfc column to add z coordinates to.}

\item{z_column}{A string of the name of the column containing elevation data.}

\item{new_column}{A string of the name of the modified sfc column. If NULL, sfc_column will be modified.}

\item{remove_z}{A logical indicating whether to remove the z_column after modifying sfc_column.}
}
\value{
The modified object.
}
\description{
sfc_column and z_column must have identical number of rows for binding. z_column of 'units' class is accepted.
If new_column is NULL, sfc_column will be replaced with the modifed sfc column.
}
