% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{ps_sfc_centroid1}
\alias{ps_sfc_centroid1}
\title{Find centroid of POINT sfc}
\usage{
ps_sfc_centroid1(x, sfc_name = ps_active_sfc_name(x), by = character(0),
  nearest = FALSE)
}
\arguments{
\item{x}{The object}

\item{sfc_name}{A string of the sfc column name}

\item{by}{A character vector of the columns to get the centroid by.}

\item{nearest}{A flag indicating whether to return the point closest to the centroid (as opposed to the actual centroid)}
}
\value{
Sf object of centroid
}
\description{
Find centroid of POINT sfc
}
