% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_sfc_name}
\alias{ps_sfc_name}
\title{Name of sfc column(s)}
\usage{
ps_sfc_name(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A character vector of the name of the geometry column(s) ('sfc column(s)).
}
\description{
Name of sfc column(s)
}
