% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_sfc_names}
\alias{ps_sfc_names}
\title{Name of sfc column(s)}
\usage{
ps_sfc_names(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A character vector of the name of the sfc (inactive geometry) column(s).
}
\description{
Name of sfc column(s)
}
