% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_sfc_rectangle}
\alias{ps_sfc_rectangle}
\title{Create sfc rectangle}
\usage{
ps_sfc_rectangle(x, ...)
}
\arguments{
\item{x}{A bbox object (or numeric vector of length 4 indicating xmin, ymin, xmax, ymax).}

\item{...}{zero or more simple feature geometries (objects of class \code{sfg}), or a single list of such objects; \code{NULL} values will get replaced by empty geometries.}
}
\value{
sfc polygon.
}
\description{
Create sfc rectangle
}
