% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlat.R
\name{ps_sfc_to_longlat}
\alias{ps_sfc_to_longlat}
\title{Convert sfc column to Longitude and Latitude in WGS84.}
\usage{
ps_sfc_to_longlat(x, sfc_name = ps_active_sfc_name(x))
}
\arguments{
\item{x}{The sf object.}

\item{sfc_name}{A string of the sfc column name.}
}
\value{
A tibble with Longitude and Latitude.
}
\description{
Convert sfc column to Longitude and Latitude in WGS84.
}
