% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat.R
\name{ps_sfc_to_lonlat}
\alias{ps_sfc_to_lonlat}
\title{Convert sf (active geometry) column to Longitude and Latitude in WGS84.}
\usage{
ps_sfc_to_lonlat(x, sfc_name)
}
\arguments{
\item{x}{The sf object.}

\item{sfc_name}{A string of the sfc column name.}
}
\value{
A tibble with Longitude and Latitude.
}
\description{
Convert sf (active geometry) column to Longitude and Latitude in WGS84.
}
