% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{ps_sfcs_centroid}
\alias{ps_sfcs_centroid}
\title{Find centroid of sfcs}
\usage{
ps_sfcs_centroid(x = sf, sfc_names = ps_sfc_names(x))
}
\arguments{
\item{x}{The object}

\item{sfc_names}{A character vector of the sfc column names}
}
\value{
Sf object of centroid
}
\description{
Centroid can be found for multiple sfc columns and any geometry type.
For polygon and linestring data, centroid is calculated from vertices.
Sfc columns must have same crs and should not be longitude/latitude.
}
