% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{ps_sfcs_to_crs}
\alias{ps_sfcs_to_crs}
\title{Reproject sfc columns}
\usage{
ps_sfcs_to_crs(x, sfc_names = ps_sfc_names(x), crs = getOption("ps.crs", 4326))
}
\arguments{
\item{x}{The object}

\item{sfc_names}{A character vector of the sfc column names}

\item{crs}{The projection to use.}
}
\value{
The modified object
}
\description{
Reproject sfc columns
}
