% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{ps_sfcs_to_utm}
\alias{ps_sfcs_to_utm}
\title{Reproject sfcs column to UTMs.}
\usage{
ps_sfcs_to_utm(x, sfc_names = ps_sfc_names(x))
}
\arguments{
\item{x}{The object}

\item{sfc_names}{A character vector of the sfc column names}
}
\value{
The modified object
}
\description{
The UTM zone is that for the centroid of the centroids of the column.
}
