% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{ps_sfcs_to_wgs84}
\alias{ps_sfcs_to_wgs84}
\title{Reproject sfc columns to WGS84}
\usage{
ps_sfcs_to_wgs84(x, sfc_names = ps_sfc_names(x))
}
\arguments{
\item{x}{The object}

\item{sfc_names}{A character vector of the sfc column names}
}
\value{
The modified object
}
\description{
Reproject sfc columns to WGS84
}
