% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{ps_sfg_rectangle}
\alias{ps_sfg_rectangle}
\title{Create sfg rectangle}
\usage{
ps_sfg_rectangle(x)
}
\arguments{
\item{x}{A bbox object (or numeric vector of length 4 indicating xmin, ymin, xmax, ymax).}
}
\value{
sfg polygon.
}
\description{
Create sfg rectangle
}
