% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ps_subset_fwa}
\alias{ps_subset_fwa}
\title{Filter/clip a FWA dataset.}
\usage{
ps_subset_fwa(x, ws_codes = c("BULL", "CLRH", "COLR", "DUNC", "ELKR", "KHOR",
  "KOTL", "KOTR", "LARL", "REVL", "SLOC", "SMAR", "UARL"), clip = NULL)
}
\arguments{
\item{x}{A sf object to subset.}

\item{ws_codes}{A character vector specifying the watershed codes.}

\item{clip}{A sf/sfc object to clip dataset to.}
}
\value{
sf object.
}
\description{
If ws_code argument not used, this is simply a wrapper on st_intersection.
}
