% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ps_ws_codes}
\alias{ps_ws_codes}
\title{Watershed Codes}
\usage{
ps_ws_codes()
}
\value{
A factor of the 13 watershed codes for the Kootenay Region of British Columbia.
}
\description{
Watershed Codes
}
\examples{
ps_ws_codes()
}
