% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{is_longlat_real}
\alias{is_longlat_real}
\title{Test if object without crs is long/lat (longitude between 180 and -180). Object assumed to be in format c(long, lat, long, lat) as in bbox.}
\usage{
is_longlat_real(x)
}
\arguments{
\item{x}{object to test}
}
\value{
flag
}
\description{
Test if object without crs is long/lat (longitude between 180 and -180). Object assumed to be in format c(long, lat, long, lat) as in bbox.
}
