% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{ps_coords_to_sfc}
\alias{ps_coords_to_sfc}
\title{Convert point coordinates to sfc column.}
\usage{
ps_coords_to_sfc(
  x,
  coords = c("X", "Y"),
  crs = getOption("ps.crs", 4326),
  sfc_name = "geometry",
  activate = TRUE
)
}
\arguments{
\item{x}{The object.}

\item{coords}{A character vector of specifying the two columns with the point information.}

\item{crs}{An integer with the EPSG code, or character with proj4string.}

\item{sfc_name}{A string of the name of the sfc column to create.}

\item{activate}{A flag indicating whether to activate the sfc.}
}
\value{
The modified object with the coordinates removed
}
\description{
If sfc_name is the active then the new sfc column replaces it and becomes the
active one.
}
