% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{ps_nearest}
\alias{ps_nearest}
\title{Nearest}
\usage{
ps_nearest(x, y, by)
}
\arguments{
\item{x}{A data.frame.}

\item{y}{A data.frame.}

\item{by}{The unquoted name of a shared column.}
}
\description{
Nearest
}
