% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{ps_nearest}
\alias{ps_nearest}
\title{Nearest Join}
\usage{
ps_nearest(x, y, by = c("X", "Y"), dist_col = NULL, ...)
}
\arguments{
\item{x}{The object.}

\item{y}{The object with rows to bind to each row in y.}

\item{by}{A character vector of length 1 or 2 specifying the columns to use when calculating distance.}

\item{dist_col}{A string indicating the name of the column to save the distance in.}

\item{...}{Not used}
}
\value{
The updated object with the columns in y.
}
\description{
Each row in object x is bound with a closet row in object y based on the closest values in by.
}
