% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{ps_nearest}
\alias{ps_nearest}
\title{Nearest}
\usage{
ps_nearest(x, y, by)
}
\arguments{
\item{x}{A data.frame with the column by}

\item{y}{A data.frame.}

\item{by}{The unquoted name of the column to calculate the inner join by.}
}
\value{
xx
}
\description{
Inner join on two data.frames based on the nearest value in by.
}
