% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ps_read_fwa}
\alias{ps_read_fwa}
\title{Read a FWA layer.}
\usage{
ps_read_fwa(layer, path = "~/Poisson/Data/spatial/fwa/")
}
\arguments{
\item{layer}{A character string indicating which dataset to read. Options are "lakes", "rivers", "watersheds", "riverpoly"}

\item{path}{A character string indicating path to FWA data.}
}
\value{
sf object.
}
\description{
Internal function that uses shared Poisson dropbox folder.
}
