% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_remove_sfcs}
\alias{ps_remove_sfcs}
\title{Remove sfc columns}
\usage{
ps_remove_sfcs(x, sfc_columns = ps_sfc_names(x))
}
\arguments{
\item{x}{An sf object.}

\item{sfc_columns}{A character vector indicating the name of the sfc column(s) to remove.}
}
\description{
Remove sfc (inactive geometry) column(s). Active sfc column ('sf column') will not be removed even if specified.
}
