% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_set_sf}
\alias{ps_set_sf}
\title{Set sf column}
\usage{
ps_set_sf(x, sfc_column = "geometry")
}
\arguments{
\item{x}{A sf object.}

\item{sfc_column}{A string indicating the name of the sf column to set.}
}
\description{
Set active geometry column ('sf column').
}
\details{
This is a wrapper for sf::st_set_geometry().
}
