% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-geometry-column.R
\name{ps_sfc_to_coords}
\alias{ps_sfc_to_coords}
\title{Convert sfc (inactive geometry) to pair of coordinates column.}
\usage{
ps_sfc_to_coords(x, sfc_name = "geometry", X = "X", Y = "Y")
}
\arguments{
\item{x}{The object with columns}

\item{sfc_name}{A string of the sfc name.}

\item{X}{A string of the name of the X coordinate.}

\item{Y}{A string of the name of the Y coordinate.}
}
\value{
The modified object with the sfc column removed
}
\description{
Convert sfc (inactive geometry) to pair of coordinates column.
}
